Amending function to more accurately check a channel's status
Twitch appears to change <div> element object names, and such is an unreliable method for determining a channel's status. As a result, that has been deprioritized, but left in place as a fallback.

The function has been changed to look for <script> elements wherein, if a channel is live, a JSON object is present therein that contains ""isLiveBroadcast":true". If such is present, then the channel is online. Likewise, a check for the element "@type":"VideoObject" is also present. In testing, this does not appear to produce false positives for channels that are hosting.